Add delayed injection loading
This commit causes the data to load on the first sign of the page
changing, but waits until the page is loaded enough before injecting.
This means injection is fast, data loading is undetectable, and our
comments will display in synch with youtube comments. Code is pretty
hack jobby, so a basic rundown is. Load data and ready injection are
called to mark when injection is possible and when data has been loaded.
They then both call a handler which will only run the main script if
both the data is loaded and the page is ready for injection. There are
some places where the variables used to track this are reset to false.
They may appear abitrary but they have a purpose i promise ;)